Example of post-deploy hook doing rhel registration
Illustrates one potential usage of the post-deploy hooks to register
a server post deploy with RHN or a satellite.

Note this requires some additional parameters, which can be specified in
environment_rhel_reg.yaml, and this must be passed into the call to heat
via another -e parameter.  An alternative may be to have a global
extraconfig_env.yaml at the top level, which the scripts always pass, or
to use the global environment (/etc/heat/environment.d/default.yaml) on
the seed.

Change-Id: Ia6fd270122cbc2e51beb672654e5e1ebd3bd2966