Introduce Overcloud Log Aggregation
This commit introduces overcloud log aggregation to puppet-based
installs, using fluentd as a collection agent.

The requisite packages and puppet modules should already
have been installed by the 'overcloud-logging' element in
tripleo-puppet-elements.

The installed fluentd instance will forward collected logs to the
fluentd instance on the given log aggregation server.  To enable this,
use the `LogAggregationServer` parameter on the compute and controller
templates (either through a Tuskar plan or the
'overcloud-without-mergepy.yaml' template).  This parameter should contain
the IP address or hostname of the main log aggregation server
(e.g. the undercloud).

Depends-On: I87524bc81621d7a1c4a1ab8f7c6ac049b1443371
Change-Id: Ic0af2ca635e7dad27945fcbee2df8afe2c72a83c

Conflicts:
	puppet/hieradata/common.yaml