Config & deployments to update overcloud packages
This change adds config and deployment resources to trigger package
updates on nodes. The deployments are triggered by doing a stack-update
and setting one of the parameters to a unique value.

The intent is that rolling update will be controlled by setting
breakpoints on all of the UpdateDeployment resources inside the
role resource groups.

Change-Id: I56bbf944ecd6cbdbf116021b8a53f9f9111c134f

Conflicts:
	overcloud-resource-registry.yaml
	puppet/compute-puppet.yaml
	puppet/controller-puppet.yaml